Rename CombinationTouchListener to TouchListenerCollection
